@@ -4,7 +4,7 @@ publish_to: none
 version: '0.4.0'
 
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   path: any
@@ -17,7 +17,7 @@ dependencies:
   matcher:
   meta:
 dev_dependencies:
-  flutter_lints:
+  flutter_lints: '>=2.0.0'
   flutter_test:
     sdk: flutter
   integration_test:
