@@ -3,13 +3,26 @@ repository: https://github.com/tekartik/sqflite/tree/master/sqflite
 issue_tracker: https://github.com/tekartik/sqflite/issues
 description: Flutter plugin for SQLite, a self-contained, high-reliability,
   embedded, SQL database engine.
-version: 2.2.6
+version: 2.3.3+1
+
+# This package supports all platforms listed below.
+platforms:
+  android:
+  ios:
+  macos:
+
+topics:
+  - sql
+  - database
+
 funding:
   - https://github.com/sponsors/alextekartik
 
 environment:
-  sdk: '>=2.18.0 <3.0.0'
-  flutter: ">=3.3.0"
+  sdk: '>=3.3.0 <4.0.0'
+  # Flutter versions prior to 3.7 did not support the
+  # sharedDarwinSource option.
+  flutter: ">=3.7.0"
 
 flutter:
   plugin:
@@ -17,15 +30,20 @@ flutter:
       android:
         package: com.tekartik.sqflite
         pluginClass: SqflitePlugin
+        dartPluginClass: SqflitePlugin
       ios:
         pluginClass: SqflitePlugin
+        dartPluginClass: SqflitePlugin
+        sharedDarwinSource: true
       macos:
         pluginClass: SqflitePlugin
+        dartPluginClass: SqflitePlugin
+        sharedDarwinSource: true
 
 dependencies:
   flutter:
     sdk: flutter
-  sqflite_common: '>=2.4.3 <4.0.0'
+  sqflite_common: '>=2.5.0-1 <4.0.0'
   path: '>=1.8.0 <3.0.0'
 
 dev_dependencies:
