@@ -1,25 +1,36 @@
 name: sqflite_common_ffi
 homepage: https://github.com/tekartik/sqflite/tree/master/sqflite_common_ffi
 description: sqflite ffi based implementation, for desktop and units tests.
-version: 2.2.2
+version: 2.3.3
 funding:
   - https://github.com/sponsors/alextekartik
 
+# This package supports all platforms listed below.
+platforms:
+  linux:
+  macos:
+  windows:
+  android:
+  ios:
+
+topics:
+  - sql
+  - database
+
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
-  sqlite3: '>=1.9.1 <3.0.0'
-  sqflite_common: '>=2.4.3 <4.0.0'
+  sqlite3: '>=2.3.0 <4.0.0'
+  sqflite_common: '>=2.5.4 <4.0.0'
   synchronized: '>=3.0.0 <5.0.0'
   path: '>=1.8.0 <3.0.0'
   meta: '>=1.3.0 <3.0.0'
 
 dev_dependencies:
-  lints:
+  lints: '>=2.1.1'
   test: '>=1.16.2'
   process_run: '>=0.12.0'
   pub_semver: '>=2.0.0'
-  archive:
+  archive: '>=3.5.0'
   http:
-
