@@ -5,12 +5,12 @@ homepage: https://github.com/tekartik/sqflite
 publish_to: none
 
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dev_dependencies:
   lints: '>=1.0.0'
   process_run: '>=0.11.0+2'
-  dev_test: '>=0.13.4-dev.2'
+  dev_build: '>=0.16.2'
   test:
   path:
   pub_semver:
