@@ -1,6 +1,6 @@
 import 'dart:async';
-import 'dart:html' as html;
 import 'dart:math';
+import 'dart:typed_data';
 
 import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common/utils/utils.dart';
@@ -8,6 +8,7 @@ import 'package:sqflite_common_ffi_web/sqflite_ffi_web.dart';
 import 'package:sqflite_common_ffi_web/src/sw/constants.dart';
 import 'package:sqflite_common_ffi_web/src/web/load_sqlite_web.dart'
     show SqfliteFfiWebContextExt;
+import 'package:web/web.dart' as web;
 
 import 'ui.dart';
 
@@ -21,7 +22,10 @@ var swOptions = SqfliteFfiWebOptions(
     sharedWorkerUri: Uri.parse('sw.dart.js'),
     // ignore: invalid_use_of_visible_for_testing_member
     forceAsBasicWorker: _useBasicWebWorker);
-
+var swBasicOptions = SqfliteFfiWebOptions(
+    sharedWorkerUri: Uri.parse('sw.dart.js'),
+    // ignore: invalid_use_of_visible_for_testing_member
+    forceAsBasicWorker: true);
 Future incrementPrebuilt() async {
   await incrementSqfliteValueInDatabaseFactory(databaseFactoryWebPrebuilt,
       tag: 'prebuilt');
@@ -110,8 +114,9 @@ Future incrementNoWebWorker() async {
 }
 
 Future<void> main() async {
+  // sqliteFfiWebDebugWebWorker = true;
   initUi();
-  // sqliteFfiWebDebugWebWorker = devWarning(true);
+
   write('$_shc running $_debugVersion');
   // devWarning(incrementVarInSharedWorker());
   // await devWarning(bigInt());
@@ -128,15 +133,22 @@ Future<void> main() async {
 
 var _webContextRegisterAndReady = sqfliteFfiWebStartSharedWorker(swOptions);
 
-Future<html.SharedWorker> sharedWorkerRegisterAndReady() async =>
+var _webBasicContextRegisterAndReady =
+    sqfliteFfiWebStartSharedWorker(swBasicOptions);
+
+Future<web.SharedWorker> sharedWorkerRegisterAndReady() async =>
     (await _webContextRegisterAndReady).sharedWorker!;
 
 Future<SqfliteFfiWebContext> webContextRegisterAndReady() async =>
     (await _webContextRegisterAndReady);
 
+Future<SqfliteFfiWebContext> webBasicContextRegisterAndReady() async =>
+    (await _webBasicContextRegisterAndReady);
 var databaseFactoryWebPrebuilt = databaseFactoryFfiWeb;
 var databaseFactoryWebNoWebWorkerLocal = databaseFactoryFfiWebNoWebWorker;
 var databaseFactoryWebLocal = createDatabaseFactoryFfiWeb(options: swOptions);
+var databaseFactoryWebBasicWorkerLocal =
+    createDatabaseFactoryFfiWeb(options: swBasicOptions);
 
 var key = 'testValue';
 
@@ -158,6 +170,20 @@ Future<void> setTestValue(SqfliteFfiWebContext context, Object? value) async {
 Future<void> incrementVarInSharedWorker() async {
   var context = await webContextRegisterAndReady();
   write('shared worker ready');
+  var value = await getTestValue(context);
+  write('var before $value');
+  if (value is! int) {
+    value = 0;
+  }
+
+  await setTestValue(context, value + 1);
+  value = await getTestValue(context);
+  write('var after $value');
+}
+
+Future<void> incrementVarInBasicdWorker() async {
+  var context = await webBasicContextRegisterAndReady();
+  write('basic worker ready');
   var value = await getTestValue(context);
   write('var before $value');
   if (value is! int) {
@@ -202,10 +228,25 @@ Future<void> incrementSqfliteValueInDatabaseFactory(DatabaseFactory factory,
   }
 }
 
+Future<void> readWiteDatabase(DatabaseFactory factory, int size) async {
+  try {
+    var path = 'read_write.db';
+    var bytes = Uint8List.fromList(List.generate(size, (index) => index % 256));
+    await factory.writeDatabaseBytes(path, bytes);
+    var readBytes = await factory.readDatabaseBytes(path);
+    write('wrote ${bytes.length}, read ${readBytes.length}');
+  } catch (e) {
+    write('Exception $e');
+  }
+}
+
 void initUi() {
   addButton('load sqlite', () async {});
   addButton('increment var in shared worker', () async {
     await incrementVarInSharedWorker();
+  });
+  addButton('increment var in basic worker', () async {
+    await incrementVarInBasicdWorker();
   });
   addButton('increment sqflite value in main thread', () async {
     await incrementNoWebWorker();
@@ -218,5 +259,16 @@ void initUi() {
   });
   addButton('increment sqflite value in pre-built web worker', () async {
     await incrementPrebuilt();
+  });
+  addButton('read write file', () async {
+    for (var factory in [
+      databaseFactoryWebNoWebWorkerLocal,
+      databaseFactoryWebPrebuilt,
+      databaseFactoryWebLocal
+    ]) {
+      write('factory: $factory');
+      await readWiteDatabase(factory, 3);
+      await readWiteDatabase(factory, 1024 * 1024);
+    }
   });
 }
