@@ -1,30 +1,30 @@
-import 'dart:html';
-import 'dart:js_util';
-import 'dart:typed_data';
+import 'dart:js_interop';
 
-import 'package:http/http.dart' as http;
 import 'package:sqflite_common_ffi_web/sqflite_ffi_web.dart';
 import 'package:sqflite_common_ffi_web/src/constant.dart';
-import 'package:sqflite_common_ffi_web/src/debug/debug.dart';
-import 'package:sqflite_common_ffi_web/src/sqflite_ffi_impl_web.dart';
 import 'package:sqlite3/wasm.dart';
+import 'package:web/web.dart' as web;
+
+import 'worker_message_utils.dart';
 
 bool get _debug => sqliteFfiWebDebugWebWorker;
 
 var _swc = workerClientLogPrefix; // Log prefix
+var _log = print;
 
 /// Load base file system
 Future<SqfliteFfiWebContext> sqfliteFfiWebLoadSqlite3FileSystem(
     SqfliteFfiWebOptions options) async {
-  // devPrint('options');
   var indexedDbName = options.indexedDbName ?? 'sqflite_databases';
   final fs = await IndexedDbFileSystem.open(dbName: indexedDbName);
   return SqfliteFfiWebContextImpl(options: options, fs: fs);
 }
 
 var _defaultSqlite3WasmUri = Uri.parse('sqlite3.wasm');
-var _defaultSharedWorkerUri = Uri.parse(sqfliteSharedWorkerJsFile);
-//var uri = options.sqlite3WasmUri ?? _defaultSqlite3WasmUri;
+
+/// Default shared worker Uri.
+var defaultSharedWorkerUri = Uri.parse(sqfliteSharedWorkerJsFile);
+
 /// Default indexedDB name is /sqflite
 Future<SqfliteFfiWebContext> sqfliteFfiWebLoadSqlite3Wasm(
     SqfliteFfiWebOptions options,
@@ -32,25 +32,15 @@ Future<SqfliteFfiWebContext> sqfliteFfiWebLoadSqlite3Wasm(
     bool? fromWebWorker}) async {
   context ??= await sqfliteFfiWebLoadSqlite3FileSystem(options);
   var uri = options.sqlite3WasmUri ?? _defaultSqlite3WasmUri;
-  Uint8List bodyBytes;
   if (_debug) {
-    print('Loading sqlite3.wasm from $uri');
+    _log('Loading sqlite3.wasm from $uri');
   }
-  if (fromWebWorker ?? false) {
-    var self = WorkerGlobalScope.instance;
-    final response = (await self.fetch(uri.toString())) as Object;
-    bodyBytes = ((await promiseToFuture<Object?>(
-            callMethod(response, 'arrayBuffer', []))) as ByteBuffer)
-        .asUint8List();
-  } else {
-    // regular http
-    final response = await http.get(uri);
-    bodyBytes = response.bodyBytes;
-  }
+
   var webContext = (context as SqfliteFfiWebContextImpl);
-  final fs = webContext.fs;
-  var wasmSqlite3 =
-      await WasmSqlite3.load(bodyBytes, SqliteEnvironment(fileSystem: fs));
+  final fs = webContext.fs ?? InMemoryFileSystem();
+  var wasmSqlite3 = await WasmSqlite3.loadFromUrl(uri);
+  wasmSqlite3.registerVirtualFileSystem(fs, makeDefault: true);
+
   return SqfliteFfiWebContextImpl(
       options: options, fs: fs, wasmSqlite3: wasmSqlite3);
 }
@@ -60,34 +50,34 @@ Future<SqfliteFfiWebContext> sqfliteFfiWebStartSharedWorker(
     SqfliteFfiWebOptions options) async {
   try {
     var name = 'sqflite_common_ffi_web';
-    var sharedWorkerUri = options.sharedWorkerUri ?? _defaultSharedWorkerUri;
-    SharedWorker? sharedWorker;
-    Worker? worker;
+    var sharedWorkerUri = options.sharedWorkerUri ?? defaultSharedWorkerUri;
+    web.SharedWorker? sharedWorker;
+    web.Worker? worker;
     try {
       if (!(options.forceAsBasicWorker ?? false)) {
         if (_debug) {
-          print(
+          _log(
               '$_swc registering shared worker $sharedWorkerUri (name: $name)');
         }
-        sharedWorker = SharedWorker(sharedWorkerUri.toString(), name);
+        sharedWorker = web.SharedWorker(sharedWorkerUri.toString(), name.toJS);
       }
     } catch (e) {
       if (_debug) {
-        print('SharedWorker creation failed $e');
+        _log('SharedWorker creation failed $e');
       }
     }
     if (sharedWorker == null) {
       if (_debug) {
-        print('$_swc registering worker $sharedWorkerUri');
+        _log('$_swc registering worker $sharedWorkerUri');
       }
-      worker = Worker(sharedWorkerUri.toString());
+      worker = web.Worker(sharedWorkerUri.toString());
     }
     return SqfliteFfiWebContextImpl(
         options: options, sharedWorker: sharedWorker, worker: worker);
   } catch (e, st) {
     if (_debug) {
-      print(e);
-      print(st);
+      _log('sqfliteFfiWebLoadSqlite3Wasm failed: $e');
+      _log(st);
     }
     rethrow;
   }
@@ -96,28 +86,27 @@ Future<SqfliteFfiWebContext> sqfliteFfiWebStartSharedWorker(
 /// Web implementation with shared worker
 class SqfliteFfiWebContextImpl extends SqfliteFfiWebContext {
   /// Null when using shared worker
-  final FileSystem? fs;
+  final VirtualFileSystem? fs;
 
   /// Null when using shared worker
   final WasmSqlite3? wasmSqlite3;
 
   /// Optional Client shared worker
-  final SharedWorker? sharedWorker;
+  final web.SharedWorker? sharedWorker;
 
   /// Optional Client basic worker (if sharedWorker not working)
-  final Worker? worker;
+  final web.Worker? worker;
 
   /// Raw message sender to either shared worker or basic worker
   late final RawMessageSender rawMessageSender;
 
   /// Web implementation with shared worker
   SqfliteFfiWebContextImpl(
-      {required SqfliteFfiWebOptions options,
+      {required super.options,
       this.fs,
       this.wasmSqlite3,
       this.sharedWorker,
-      this.worker})
-      : super(options: options) {
+      this.worker}) {
     if (sharedWorker != null) {
       rawMessageSender = RawMessageSenderSharedWorker(sharedWorker!);
     }
@@ -132,13 +121,13 @@ extension SqfliteFfiWebContextExt on SqfliteFfiWebContext {
   SqfliteFfiWebContextImpl get _context => this as SqfliteFfiWebContextImpl;
 
   /// File system if any
-  FileSystem? get fs => _context.fs;
+  VirtualFileSystem? get fs => _context.fs;
 
   /// Shared worker if any
-  SharedWorker? get sharedWorker => _context.sharedWorker;
+  web.SharedWorker? get sharedWorker => _context.sharedWorker;
 
   /// Web worker if any
-  SharedWorker? get webWorker => _context.sharedWorker;
+  web.SharedWorker? get webWorker => _context.sharedWorker;
 
   /// Loaded wasm if any
   WasmSqlite3? get wasmSqlite3 => _context.wasmSqlite3;
@@ -146,4 +135,10 @@ extension SqfliteFfiWebContextExt on SqfliteFfiWebContext {
   /// Send raw message to worker
   Future<Object?> sendRawMessage(Object message) =>
       _context.rawMessageSender.sendRawMessage(message);
+}
+
+/// Web specific exception. For now only sent when catching a web worker error.
+class SqfliteFfiWebWorkerException implements Exception {
+  @override
+  String toString() => 'SqfliteFfiWebException()';
 }
