@@ -1,5 +1,6 @@
 @TestOn('browser')
-import 'package:sqflite_common_ffi/sqflite_ffi.dart';
+library;
+
 import 'package:sqflite_common_ffi_web/sqflite_ffi_web.dart';
 import 'package:sqflite_common_test/all_test.dart' as all;
 import 'package:sqflite_common_test/sqflite_test.dart';
@@ -14,11 +15,19 @@ class SqfliteFfiWebNoWebWorkerTestContext extends SqfliteLocalTestContext {
 var ffiTestContext = SqfliteFfiWebNoWebWorkerTestContext();
 
 Future<void> main() async {
-  /// Initialize ffi loader
-  sqfliteFfiInit();
-  // Add _no_isolate suffix to the path
-  var dbsPath = await _factory.getDatabasesPath();
-  await _factory.setDatabasesPath('${dbsPath}_no_web_worker');
+  /// Tmp debug
+  // sqliteFfiWebDebugWebWorker = true;
 
-  all.run(ffiTestContext);
+  /// Initialize ffi loader
+  //sqfliteFfiInit();
+  // Add _no_isolate suffix to the path
+  try {
+    var dbsPath = await _factory.getDatabasesPath();
+    await _factory.setDatabasesPath('${dbsPath}_no_web_worker');
+
+    all.run(ffiTestContext);
+  } catch (e) {
+    print('Please run setup_web_tests first');
+    test('Please run setup_web_tests first', () {}, skip: true);
+  }
 }
