@@ -1,80 +1,37 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:sqflite/sqflite.dart';
 import 'package:sqflite/src/exception_impl.dart' as impl;
 import 'package:sqflite/src/sqflite_impl.dart' as impl;
 import 'package:sqflite/src/sqflite_import.dart';
+import 'package:sqflite_common/sqflite.dart' as sqflite_common;
+// ignore: implementation_imports
+import 'package:sqflite_common/src/mixin/platform.dart';
 
 import 'dev_utils.dart'; // ignore: unused_import
-
-SqfliteDatabaseFactory? _databaseFactory;
-
-/// sqflite Default factory
-DatabaseFactory get databaseFactory => sqfliteDatabaseFactory;
-
-/// Change the default factory.
-///
-/// Be aware of the potential side effect. Any library using sqflite
-/// will have this factory as the default for all operations.
-///
-/// This setter must be call only once, before any other calls to sqflite.
-set databaseFactory(DatabaseFactory? databaseFactory) {
-  // Warn when changing. might throw in the future
-  if (databaseFactory != null) {
-    if (databaseFactory is! SqfliteDatabaseFactory) {
-      throw ArgumentError.value(
-          databaseFactory, 'databaseFactory', 'Unsupported sqflite factory');
-    }
-    if (_databaseFactory != null) {
-      stderr.writeln('''
-*** sqflite warning ***
-
-You are changing sqflite default factory.
-Be aware of the potential side effects. Any library using sqflite
-will have this factory as the default for all operations.
-
-*** sqflite warning ***
-''');
-    }
-    sqfliteDatabaseFactory = databaseFactory;
-  } else {
-    /// Will use the plugin sqflite factory
-    sqfliteDatabaseFactory = null;
-  }
-}
-
-/// sqflite Default factory
-///
-/// Definition with a typo error.
-/// - Will be soon deprecated
-/// - Will be removed in 2.0.0
-@Deprecated('Use databaseFactory instead (typo error)')
-SqfliteDatabaseFactory get sqlfliteDatabaseFactory => sqfliteDatabaseFactory;
-
-/// Change the default factory. test only.
-///
-/// Definition with a typo error.
-///
-/// Will be removed in 2.0.0
-@Deprecated('Use databaseFactory')
-set sqlfliteDatabaseFactory(SqfliteDatabaseFactory? databaseFactory) =>
-    _databaseFactory = databaseFactory;
 
 /// sqflite Default factory
 @visibleForTesting
 SqfliteDatabaseFactory get sqfliteDatabaseFactory =>
-    _databaseFactory ??= sqfliteDatabaseFactoryDefault;
+    // ignore: invalid_use_of_visible_for_testing_member
+    (databaseFactoryOrNull ?? databaseFactorySqflitePlugin)
+        as SqfliteDatabaseFactory;
+
+final SqfliteDatabaseFactory _databaseFactorySqflitePlugin =
+    SqfliteDatabaseFactoryImpl();
 
 /// Default factory that uses the plugin.
-SqfliteDatabaseFactory sqfliteDatabaseFactoryDefault =
-    SqfliteDatabaseFactoryImpl();
+DatabaseFactory get databaseFactorySqflitePlugin =>
+    _databaseFactorySqflitePlugin;
+
+/// Default factory that uses the plugin.
+final sqfliteDatabaseFactoryDefault = _databaseFactorySqflitePlugin;
 
 /// Change the default factory. test only.
 @visibleForTesting
 set sqfliteDatabaseFactory(SqfliteDatabaseFactory? databaseFactory) =>
-    _databaseFactory = databaseFactory;
+    sqflite_common.databaseFactory = databaseFactory;
 
 /// Factory implementation
 class SqfliteDatabaseFactoryImpl with SqfliteDatabaseFactoryMixin {
@@ -99,5 +56,15 @@ class SqfliteDatabaseFactoryImpl with SqfliteDatabaseFactoryMixin {
     // ignore: avoid_print
     print('<- $result');
     return result;
+  }
+
+  @override
+  Future<Uint8List> readDatabaseBytes(String path) async {
+    return await platform.databaseFileSystem.readDatabaseBytes(path);
+  }
+
+  @override
+  Future<void> writeDatabaseBytes(String path, Uint8List bytes) async {
+    await platform.databaseFileSystem.writeDatabaseBytes(path, bytes);
   }
 }
