@@ -4,8 +4,10 @@ import 'package:sqflite/sqflite.dart';
 
 import 'src_mixin_test.dart' show MockDatabaseFactoryEmpty, MockInvalidFactory;
 
+T? _ambiguate<T>(T? value) => value;
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
+  databaseFactory = databaseFactorySqflitePlugin;
 
   group('sqflite', () {
     const channel = MethodChannel('com.tekartik.sqflite');
@@ -13,7 +15,9 @@ void main() {
     final log = <MethodCall>[];
     String? response;
 
-    channel.setMockMethodCallHandler((MethodCall methodCall) async {
+    _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+        .defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, (MethodCall methodCall) async {
       log.add(methodCall);
       return response;
     });
@@ -23,6 +27,7 @@ void main() {
     });
 
     test('setDebugModeOn', () async {
+      // ignore: deprecated_member_use_from_same_package
       await Sqflite.setDebugModeOn();
       expect(log.first.method, 'debugMode');
       expect(log.first.arguments, true);
