@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:typed_data';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:path/path.dart';
@@ -34,8 +35,7 @@ dynamic mockResult(String method) {
 }
 
 class MockDatabase extends SqfliteDatabaseBase {
-  MockDatabase(SqfliteDatabaseOpenHelper openHelper, [String name = 'mock'])
-      : super(openHelper, name);
+  MockDatabase(super.openHelper, [super.name = 'mock']);
 
   int? version;
   List<String> methods = <String>[];
@@ -149,6 +149,16 @@ class MockInvalidFactory extends DatabaseFactory {
 
   @override
   Future<void> setDatabasesPath(String path) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<Uint8List> readDatabaseBytes(String path) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<void> writeDatabaseBytes(String path, Uint8List bytes) {
     throw UnimplementedError();
   }
 }
