@@ -9,15 +9,18 @@ import 'package:sqflite_common/src/factory.dart';
 import 'package:sqflite_common/src/path_utils.dart';
 import 'package:sqflite_common/src/sql_builder.dart';
 import 'package:sqflite_common/src/transaction.dart';
-import 'package:sqflite_common/src/utils.dart';
 import 'package:sqflite_common/src/utils.dart' as utils;
+import 'package:sqflite_common/src/utils.dart';
 import 'package:sqflite_common/src/value_utils.dart';
 import 'package:sqflite_common/utils/utils.dart';
 import 'package:synchronized/synchronized.dart';
 
 /// Base database implementation
 class SqfliteDatabaseBase
-    with SqfliteDatabaseMixin, SqfliteDatabaseExecutorMixin {
+    with
+        SqfliteDatabaseMixin,
+        SqfliteDatabaseWithOpenHelperMixin,
+        SqfliteDatabaseExecutorMixin {
   /// ctor
   SqfliteDatabaseBase(SqfliteDatabaseOpenHelper openHelper, String path,
       {OpenDatabaseOptions? options}) {
@@ -305,7 +308,7 @@ extension SqfliteDatabaseMixinExt on SqfliteDatabase {
   SqfliteDatabaseMixin get _mixin => this as SqfliteDatabaseMixin;
 
   /// try if open in read-only mode.
-  bool get readOnly => _mixin.openHelper?.options?.readOnly ?? false;
+  bool get readOnly => _mixin.options?.readOnly ?? false;
 
   /// for Update sql query
   /// returns the update count
@@ -324,7 +327,7 @@ extension SqfliteDatabaseMixinExt on SqfliteDatabase {
   Future<T> _txnTransaction<T>(
       Transaction? txn, Future<T> Function(Transaction txn) action,
       {bool? exclusive}) async {
-    bool? successfull;
+    bool? successful;
     var transactionStarted = txn == null;
     if (transactionStarted) {
       txn = await beginTransaction(exclusive: exclusive);
@@ -332,11 +335,13 @@ extension SqfliteDatabaseMixinExt on SqfliteDatabase {
     T result;
     try {
       result = await action(txn);
-      successfull = true;
+      successful = true;
+    } on SqfliteTransactionRollbackSuccess<T> catch (e) {
+      result = e.result;
     } finally {
       if (transactionStarted) {
         final sqfliteTransaction = txn as SqfliteTransaction;
-        sqfliteTransaction.successful = successfull;
+        sqfliteTransaction.successful = successful;
         await endTransaction(sqfliteTransaction);
       }
     }
@@ -348,7 +353,7 @@ extension SqfliteDatabaseMixinExt on SqfliteDatabase {
       {bool? exclusive}) async {
     Object? response;
     // never create transaction in read-only mode
-    if (readOnly != true) {
+    if (!readOnly) {
       if (exclusive == true) {
         response = await txnExecute<dynamic>(txn, 'BEGIN EXCLUSIVE', null,
             beginTransaction: true);
@@ -368,23 +373,36 @@ extension SqfliteDatabaseMixinExt on SqfliteDatabase {
 }
 
 /// Sqflite database mixin.
+mixin SqfliteDatabaseWithOpenHelperMixin implements SqfliteDatabase {
+  /// Keep our open helper for proper closing.
+  SqfliteDatabaseOpenHelper? openHelper;
+  @override
+  OpenDatabaseOptions? options;
+
+  @override
+  bool get isOpen => openHelper!.isOpen;
+
+  /// The factory.
+  SqfliteDatabaseFactory get factory => openHelper!.factory;
+}
+
+/// Sqflite database mixin.
 mixin SqfliteDatabaseMixin implements SqfliteDatabase {
   /// Invoke native method and wrap exception.
   Future<T> safeInvokeMethod<T>(String method, [Object? arguments]) =>
-      factory.wrapDatabaseException(() => invokeMethod(method, arguments));
+      safeAction<T>(() => invokeMethod<T>(method, arguments));
+
+  /// Invoke native action and wrap exception.
+  Future<T> safeAction<T>(Future<T> Function() action) =>
+      factory.wrapDatabaseException<T>(action);
 
   /// Invoke the native method of the factory.
   @override
   Future<T> invokeMethod<T>(String method, [Object? arguments]) =>
       _mixin.factory.invokeMethod(method, arguments);
 
-  /// Keep our open helper for proper closing.
-  SqfliteDatabaseOpenHelper? openHelper;
-  @override
-  OpenDatabaseOptions? options;
-
   /// The factory.
-  SqfliteDatabaseFactory get factory => openHelper!.factory;
+  SqfliteDatabaseFactory get factory;
 
   @override
   SqfliteDatabase get database => db;
@@ -397,7 +415,7 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
   bool isClosed = false;
 
   @override
-  bool get isOpen => openHelper!.isOpen;
+  bool get isOpen;
 
   @override
   late String path;
@@ -462,7 +480,16 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
       Transaction? txn, Future<T> Function(Transaction? txn) action) async {
     // If in a transaction, execute right away
     if (txn != null || doNotUseSynchronized) {
-      return await action(txn);
+      txn?.checkNotClosed();
+      try {
+        return await action(txn);
+      } on SqfliteDatabaseException catch (e) {
+        if (e.transactionClosed) {
+          // We are in a transaction, let's mark it as closed
+          txn?.markClosed();
+        }
+        rethrow;
+      }
     } else {
       // Simple timeout warning if we cannot get the lock after XX seconds
       final handleTimeoutWarning = (utils.lockWarningDuration != null &&
@@ -723,11 +750,20 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
   @override
   Future<void> endTransaction(SqfliteTransaction txn) async {
     // never commit transaction in read-only mode
-    if (readOnly != true) {
-      if (txn.successful == true) {
-        await txnExecute<dynamic>(txn, 'COMMIT', null);
+    if (!readOnly) {
+      if (txn.closed == true) {
+        // Ok done
+        inTransaction = false;
       } else {
-        await txnExecute<dynamic>(txn, 'ROLLBACK', null);
+        try {
+          if (txn.successful == true) {
+            await txnExecute<dynamic>(txn, 'COMMIT', null);
+          } else {
+            await txnExecute<dynamic>(txn, 'ROLLBACK', null);
+          }
+        } finally {
+          txn.closed = true;
+        }
       }
     }
   }
@@ -741,13 +777,21 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
     });
   }
 
+  /// Default implementation uses existing.
+  @override
+  Future<T> readTransaction<T>(Future<T> Function(Transaction txn) action) =>
+      transaction<T>((txn) async {
+        var result = await action(txn);
+        throw SqfliteTransactionRollbackSuccess(result);
+      });
+
   /// Close the database. Cannot be access anymore
   @override
   Future<void> close() => factory.closeDatabase(this);
 
   /// Close the database. Cannot be access anymore
   @override
-  Future<void> doClose() => _closeDatabase(id);
+  Future<void> doClose() => closeDatabase();
 
   @override
   String toString() {
@@ -759,7 +803,7 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
   /// id does not run any callback calls
   Future<int> openDatabase() async {
     final params = <String, Object?>{paramPath: path};
-    if (readOnly == true) {
+    if (readOnly) {
       params[paramReadOnly] = true;
     }
     // Single instance? never for standard inMemoryDatabase
@@ -786,7 +830,7 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
       // was in progress. This catches hot-restart scenario
       if (recoveredInTransaction) {
         // Don't do it for read-only
-        if (readOnly != true) {
+        if (!readOnly) {
           // We are not yet open so invoke the plugin directly
           try {
             await safeInvokeMethod<Object?>(methodExecute, <String, Object?>{
@@ -818,7 +862,7 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
         // Mark as closed now
         isClosed = true;
 
-        if (readOnly != true && inTransaction) {
+        if (!readOnly && inTransaction) {
           // Grab lock to prevent future access
           // At least we know no other request will be ran
           try {
@@ -844,14 +888,26 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
 
         // close for good
         // Catch exception, close should never fail
-        try {
-          await safeInvokeMethod<dynamic>(
-              methodCloseDatabase, <String, Object?>{paramId: databaseId});
-        } catch (e) {
-          print('error $e closing database $databaseId');
+        if (databaseId != null) {
+          try {
+            await safeAction<dynamic>(() => invokeCloseDatabase(databaseId));
+          } catch (e) {
+            print('error $e closing database $databaseId');
+          }
         }
       }
     });
+  }
+
+  /// Close the database with the given id.
+  Future<void> closeDatabase() async {
+    await _closeDatabase(db.id!);
+  }
+
+  /// Invoke close database.
+  Future<void> invokeCloseDatabase(int databaseId) async {
+    await safeInvokeMethod<dynamic>(
+        methodCloseDatabase, <String, Object?>{paramId: databaseId});
   }
 
   // To call during open
@@ -984,7 +1040,7 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
       return this;
     } catch (e) {
       print('error $e during open, closing...');
-      await _closeDatabase(databaseId);
+      await doClose();
       rethrow;
     } finally {
       // clean up open transaction
