@@ -1,4 +1,6 @@
-export 'platform_io.dart' if (dart.library.js) 'platform_web.dart';
+import 'package:sqflite_common/src/database_file_system.dart';
+
+export 'platform_io.dart' if (dart.library.js_interop) 'platform_web.dart';
 
 /// IO/web support
 abstract class Platform {
@@ -19,4 +21,7 @@ abstract class Platform {
 
   /// True if IO MacOS
   bool get isMacOS => false;
+
+  /// Database file system.
+  DatabaseFileSystem get databaseFileSystem;
 }
