@@ -127,8 +127,11 @@ class SqlBuilder {
     _writeClause(query, ' GROUP BY ', groupBy);
     _writeClause(query, ' HAVING ', having);
     _writeClause(query, ' ORDER BY ', orderBy);
-    if (limit != null) {
-      _writeClause(query, ' LIMIT ', limit.toString());
+    // See https://sqlite.org/lang_select.html
+    // offset cannot be specified without limit so ensure to set limit to -1
+    // if not set
+    if (limit != null || offset != null) {
+      _writeClause(query, ' LIMIT ', (limit ?? -1).toString());
     }
     if (offset != null) {
       _writeClause(query, ' OFFSET ', offset.toString());
