@@ -3,7 +3,6 @@ library sqflite_common_ffi.test.sqflite_factory_ffi_test;
 
 import 'dart:typed_data';
 
-import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 import 'package:test/test.dart';
 
@@ -11,7 +10,8 @@ void main() {
   // Set sqflite ffi support in test
   sqfliteFfiInit();
 
-  var databaseFactory = databaseFactoryFfi;
+  var databaseFactory = databaseFactoryFfi; //.debugQuickLoggerWrapper();
+
   test('simple sqflite example', () async {
     var db = await databaseFactory.openDatabase(inMemoryDatabasePath);
     expect(await db.getVersion(), 0);
@@ -57,5 +57,52 @@ void main() {
     }
 
     await db.close();
+  });
+
+  test('Defensive mode', () async {
+    // This should fail...
+    var db = await databaseFactory.openDatabase(inMemoryDatabasePath);
+
+    Future<void> createTable() async {
+      await db.execute('CREATE TABLE Test(value TEXT)');
+    }
+
+    Future<void> alterTable() async {
+      await db.update('sqlite_master', {'sql': 'CREATE TABLE Test(value BLOB)'},
+          where: 'name = \'Test\' and type = \'table\'');
+    }
+
+    try {
+      await createTable();
+      await alterTable();
+    } catch (e) {
+      print('Error update sqflite_master without protection (expected): $e');
+    } finally {
+      await db.close();
+    }
+
+    // This could fail...
+    db = await databaseFactory.openDatabase(inMemoryDatabasePath);
+    try {
+      await createTable();
+      await db.execute('PRAGMA writable_schema = ON');
+      await alterTable();
+    } catch (e) {
+      print(
+          'Error update sqflite_master (could happen without defensive mode): $e');
+    } finally {
+      await db.close();
+    }
+
+    db = await databaseFactory.openDatabase(inMemoryDatabasePath);
+    try {
+      await createTable();
+      // Workaround for iOS 14 / PR https://github.com/tekartik/sqflite/pull/1058
+      await db.execute('PRAGMA sqflite -- db_config_defensive_off');
+      await db.execute('PRAGMA writable_schema = ON');
+      await alterTable();
+    } finally {
+      await db.close();
+    }
   });
 }
