@@ -1,18 +1,16 @@
 @TestOn('vm')
+library;
+
 import 'dart:io';
 
 import 'package:archive/archive_io.dart';
-import 'package:http/http.dart';
 import 'package:path/path.dart';
 import 'package:process_run/shell.dart';
 import 'package:sqflite_common_ffi/src/windows/setup.dart';
+import 'package:sqflite_common_ffi/src/windows/sqlite3_info.dart';
 import 'package:test/test.dart';
 
-var windowsSqliteVersion = '3.38.2';
-var windowsZipSrc =
-    'https://www.sqlite.org/2022/sqlite-dll-win64-x64-3380200.zip';
-var windowsZipSha3 =
-    '9f71eec9a2c7f12602eaa2af76bd7c052e540502ae7a89dac540e10962e2fa35';
+import '../tool/windows_setup.dart';
 
 Future<String> computeSha3(String file, {String openssl = 'openssl'}) async {
   var line = (await run(
@@ -45,21 +43,9 @@ Future<String?> windowsFindOpenssl() async {
 }
 
 void main() {
-  var localZip = join('.local', basename(windowsZipSrc));
-  Future<bool> getZip() async {
-    if (!File(localZip).existsSync()) {
-      await Directory(dirname(localZip)).create(recursive: true);
-      try {
-        await File(localZip)
-            .writeAsBytes(await readBytes(Uri.parse(windowsZipSrc)));
-      } catch (e) {
-        stderr.writeln(
-            'Fail to fetch sqlite.zip version $windowsSqliteVersion at $windowsZipSrc');
-        return false;
-      }
-    }
-    return true;
-  }
+  var helper = Sqlite3DllSetupHelper(sqlite3Info);
+  var srcZip = sqlite3Info.srcZip;
+  var localZip = join('.local', basename(srcZip));
 
   group('sqlite3.dll', () {
     test('sha3', () async {
@@ -69,19 +55,19 @@ void main() {
         openssl = await windowsFindOpenssl();
       }
       if (openssl != null) {
-        if (await getZip()) {
+        if (await helper.getZip()) {
           var computed = await computeSha3(localZip, openssl: openssl);
-          expect(computed, windowsZipSha3);
+          expect(computed, sqlite3Info.sha3);
         }
       }
     });
 
     test('checkDll', () async {
       var dllPath = findWindowsDllPath()!;
-      if (await getZip()) {
+      if (await helper.getZip()) {
         final inputStream = InputFileStream(localZip);
         final archive = ZipDecoder().decodeBuffer(inputStream);
-        extractArchiveToDisk(archive, dirname(localZip));
+        await extractArchiveToDisk(archive, dirname(localZip));
 
         var downloadedDllContent =
             await File(join(dirname(localZip), 'sqlite3.dll')).readAsBytes();
