@@ -1,11 +1,11 @@
 @TestOn('vm')
+library;
+
 import 'dart:io';
 
-import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
+import 'package:sqflite_common_ffi/src/windows/sqlite3_info.dart';
 import 'package:test/test.dart';
-
-import 'sqflite_ffi_windows_setup_test.dart';
 
 void main() {
   if (Platform.isWindows) {
@@ -16,7 +16,7 @@ void main() {
         final results = await db.rawQuery('select sqlite_version()');
 
         var version = results.first.values.first;
-        expect(version, windowsSqliteVersion);
+        expect(version, sqlite3Info.version);
       });
     });
   }
