@@ -1,11 +1,15 @@
 @TestOn('vm')
+library;
+
 import 'dart:io';
 
-import 'package:dev_test/build_support.dart';
-import 'package:dev_test/src/run_ci.dart';
+import 'package:dev_build/build_support.dart';
+import 'package:dev_build/src/run_ci.dart';
 import 'package:path/path.dart';
 import 'package:process_run/shell_run.dart';
 import 'package:test/test.dart';
+
+var runningOnGithubAction = Platform.environment['GITHUB_ACTION'] != null;
 
 void main() {
   workflow();
@@ -77,8 +81,13 @@ void workflow({bool noBuild = false}) {
 
     test('build android', () async {
       await ensureCreate();
+      // if (!(runningOnGithubAction && Platform.isMacOS)) { // timeout on MacOS to fix
       await androidBuild();
-    }, timeout: const Timeout(Duration(minutes: 5)));
+      // }
+      await androidBuild();
+    },
+        timeout: Timeout(Duration(
+            minutes: (Platform.isWindows || Platform.isMacOS) ? 10 : 5)));
     test('add sqflite', () async {
       await ensureCreate();
       if (await pathPubspecAddDependency(dir, 'sqflite')) {
